placed api

+++ src/main/java/com/example/Swigatto/model/Cart.java
@@ -22,7 +22,7 @@ public class Cart {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
 
-    double cartTotal;
+    int cartTotal;
 
     @OneToOne
     @JoinColumn
